chore: cleanup deposit-swap box
